+++ DrawingLibrary/TriangleDrawer.cs
@@ -13,9 +13,9 @@ namespace DrawingLibrary
         public Shader Shader { get; set; }
         class AETData
         {
-            public float x;
-            public float antitangent;
-            public float yMax;
+            public double x;
+            public double antitangent;
+            public double yMax;
         }
         int[] sortedVerticesIndexes = { 0, 1, 2 };
         LinkedList<AETData> AET = new LinkedList<AETData>();
@@ -99,9 +99,9 @@ namespace DrawingLibrary
                 }
             }
         }
-        private float GetAntiTangent(IntVector2 from, IntVector2 to)
+        private double GetAntiTangent(IntVector2 from, IntVector2 to)
         {
-            return (float)(to.X - from.X) / (to.Y - from.Y);
+            return (double)(to.X - from.X) / (to.Y - from.Y);
         }
     }
 }
